Refactor TestCache
