+++ accio-tests/src/test/java/io/accio/testing/TestCache.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package io.accio.testing.bigquery;
+package io.accio.testing;
 
 import com.google.common.collect.ImmutableList;
 import com.google.inject.Key;
@@ -26,7 +26,6 @@ import io.accio.base.sqlrewrite.CacheRewrite;
 import io.accio.cache.CacheInfoPair;
 import io.accio.cache.TaskInfo;
 import io.accio.main.AccioMetastore;
-import io.accio.testing.AbstractCacheTest;
 import org.testng.annotations.Test;
 
 import java.sql.Connection;
@@ -39,25 +38,26 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.function.Supplier;
+import java.util.regex.Pattern;
 
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static io.accio.base.CatalogSchemaTableName.catalogSchemaTableName;
+import static io.accio.base.config.AccioConfig.DataSourceType.DUCKDB;
 import static io.accio.base.type.IntegerType.INTEGER;
 import static java.lang.String.format;
-import static java.util.Objects.requireNonNull;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatNoException;
 
 @Test(singleThreaded = true)
-public class TestCache
+public abstract class TestCache
         extends AbstractCacheTest
 {
-    private static final Function<String, String> dropTableStatement = (tableName) -> format("BEGIN TRANSACTION;DROP TABLE IF EXISTS %s;COMMIT;", tableName);
-    private final Supplier<AccioMDL> accioMDL = () -> getInstance(Key.get(AccioMetastore.class)).getAnalyzedMDL().getAccioMDL();
-    private final Supplier<DuckdbClient> duckdbClient = () -> getInstance(Key.get(DuckdbClient.class, ForCache.class));
-    private final SessionContext defaultSessionContext = SessionContext.builder()
-            .setCatalog("canner-cml")
-            .setSchema("tpch_tiny")
+    protected static final Function<String, String> dropTableStatement = (tableName) -> format("BEGIN TRANSACTION;DROP TABLE IF EXISTS %s;COMMIT;", tableName);
+    protected final Supplier<AccioMDL> accioMDL = () -> getInstance(Key.get(AccioMetastore.class)).getAnalyzedMDL().getAccioMDL();
+    protected final Supplier<DuckdbClient> duckdbClient = () -> getInstance(Key.get(DuckdbClient.class, ForCache.class));
+    protected final SessionContext defaultSessionContext = SessionContext.builder()
+            .setCatalog(getDefaultCatalog())
+            .setSchema(getDefaultSchema())
             .build();
 
     @Test
@@ -78,27 +78,21 @@ public class TestCache
                 "     o_custkey\n" +
                 "   , sum(o_totalprice) revenue\n" +
                 "   FROM\n" +
-                "     `%s.%s.%s`\n" +
-                "   GROUP BY o_custkey", "canner-cml", "tpch_tiny", "orders"));
+                "     %s.%s.%s\n" +
+                "   GROUP BY o_custkey", getDefaultCatalog(), getDefaultSchema(), "orders"));
         assertThat(duckdbResult.size()).isEqualTo(bqResult.size());
         assertThat(Arrays.deepEquals(duckdbResult.toArray(), bqResult.toArray())).isTrue();
 
         String errMsg = getDefaultCacheInfoPair("unqualified")
                 .flatMap(CacheInfoPair::getErrorMessage)
                 .orElseThrow(AssertionError::new);
-        assertThat(errMsg).matches("Failed to do cache for cacheInfo .*");
+        assertThat(errMsg).matches(Pattern.compile("Failed to do cache for cacheInfo .*", Pattern.DOTALL));
     }
 
     @Test
     public void testMetricWithoutCache()
     {
         assertThat(getDefaultCacheInfoPair("AvgRevenue")).isEmpty();
-    }
-
-    @Override
-    protected Optional<String> getAccioMDLPath()
-    {
-        return Optional.of(requireNonNull(getClass().getClassLoader().getResource("cache/cache_mdl.json")).getPath());
     }
 
     @Test
@@ -109,7 +103,7 @@ public class TestCache
                 PreparedStatement stmt = connection.prepareStatement("select custkey, revenue from Revenue limit 100");
                 ResultSet resultSet = stmt.executeQuery()) {
             resultSet.next();
-            assertThatNoException().isThrownBy(() -> resultSet.getLong("custkey"));
+            assertThatNoException().isThrownBy(() -> resultSet.getInt("custkey"));
             assertThatNoException().isThrownBy(() -> resultSet.getInt("revenue"));
             int count = 1;
 
@@ -124,9 +118,9 @@ public class TestCache
             stmt.setObject(1, 1202);
             try (ResultSet resultSet = stmt.executeQuery()) {
                 resultSet.next();
-                assertThatNoException().isThrownBy(() -> resultSet.getLong("custkey"));
+                assertThatNoException().isThrownBy(() -> resultSet.getInt("custkey"));
                 assertThatNoException().isThrownBy(() -> resultSet.getInt("revenue"));
-                assertThat(resultSet.getLong("custkey")).isEqualTo(1202L);
+                assertThat(resultSet.getInt("custkey")).isEqualTo(1202);
                 assertThat(resultSet.next()).isFalse();
             }
         }
@@ -162,7 +156,7 @@ public class TestCache
         try (ConnectorRecordIterator connectorRecordIterator = cacheManager.get().query(withParam, ImmutableList.of(new Parameter(INTEGER, 1202)))) {
             Object[] result = connectorRecordIterator.next();
             assertThat(result.length).isEqualTo(2);
-            assertThat(result[0]).isEqualTo(1202L);
+            assertThat(result[0]).isEqualTo(1202);
             assertThat(connectorRecordIterator.hasNext()).isFalse();
         }
     }
@@ -201,13 +195,13 @@ public class TestCache
         assertThat(tableNames).contains(mappingName);
 
         List<Object[]> duckdbResult = queryDuckdb(format("select * from \"%s\"", mappingName));
-        List<Object[]> bqResult = queryByMetadata("SELECT\n" +
+        List<Object[]> bqResult = queryByMetadata(format("SELECT\n" +
                 "     o_orderkey orderkey\n" +
                 "   , o_custkey custkey\n" +
                 "   , o_orderstatus orderstatus\n" +
                 "   , o_totalprice totalprice\n" +
                 "   , o_orderdate orderdate" +
-                " from `canner-cml`.tpch_tiny.orders");
+                " from %s.%s.orders", getDefaultCatalog(), getDefaultSchema()));
         assertThat(duckdbResult.size()).isEqualTo(bqResult.size());
         assertThat(Arrays.deepEquals(duckdbResult.toArray(), bqResult.toArray())).isTrue();
     }
