+++ accio-tests/src/test/java/io/accio/testing/bigquery/TestBigQueryCache.java
@@ -14,7 +14,6 @@
 
 package io.accio.testing.bigquery;
 
-import io.accio.testing.AbstractCacheTest;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
@@ -42,7 +41,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatNoException
 
 @Test(singleThreaded = true)
 public class TestBigQueryCache
-        extends AbstractCacheTest
+        extends AbstractCacheTestWithBigQuery
 {
     @Override
     protected Optional<String> getAccioMDLPath()
